Moved System Prompt to Settings Menu
