+++ src/components/ChatConfigMenu/ChatConfigMenu.tsx
@@ -13,7 +13,7 @@ import {
 } from '@components/ConfigMenu/ConfigMenu';
 
 import { ModelOptions } from '@type/chat';
-import { _defaultChatConfig, _defaultSystemMessage } from '@constants/chat';
+import { _defaultChatConfig } from '@constants/chat';
 import LinkIcon from '@icon/LinkIcon'
 
 const ChatConfigMenu = () => {
@@ -41,13 +41,7 @@ const ChatConfigPopup = ({
 }) => {
   const config = useStore.getState().defaultChatConfig;
   const setDefaultChatConfig = useStore((state) => state.setDefaultChatConfig);
-  const setDefaultSystemMessage = useStore(
-    (state) => state.setDefaultSystemMessage
-  );
 
-  const [_systemMessage, _setSystemMessage] = useState<string>(
-    useStore.getState().defaultSystemMessage
-  );
   const [_model, _setModel] = useState<ModelOptions>(config.model);
   const [_maxPromptTokens, _setMaxPromptToken] = useState<number>(config.maxPromptTokens);
   const [_maxGenerationTokens, _setMaxGenerationToken] = useState<number>(config.maxGenerationTokens);
@@ -72,7 +66,6 @@ const ChatConfigPopup = ({
       presence_penalty: _presencePenalty,
       frequency_penalty: _frequencyPenalty,
     });
-    setDefaultSystemMessage(_systemMessage);
     setIsModalOpen(false);
   };
 
@@ -84,7 +77,6 @@ const ChatConfigPopup = ({
     _setTopP(_defaultChatConfig.top_p);
     _setPresencePenalty(_defaultChatConfig.presence_penalty);
     _setFrequencyPenalty(_defaultChatConfig.frequency_penalty);
-    _setSystemMessage(_defaultSystemMessage);
   };
 
   return (
@@ -94,10 +86,6 @@ const ChatConfigPopup = ({
       handleConfirm={handleSave}
     >
       <div className='p-6 border-b border-gray-200 dark:border-gray-600 w-[90vw] max-w-full text-sm text-gray-900 dark:text-gray-300'>
-        <DefaultSystemChat
-          _systemMessage={_systemMessage}
-          _setSystemMessage={_setSystemMessage}
-        />
         <ModelSelector _model={_model} _setModel={_setModel} />
         <MaxTokenSlider
           _maxToken={_maxPromptTokens}
@@ -132,52 +120,6 @@ const ChatConfigPopup = ({
         </div>
       </div>
     </PopupModal>
-  );
-};
-
-const DefaultSystemChat = ({
-  _systemMessage,
-  _setSystemMessage,
-}: {
-  _systemMessage: string;
-  _setSystemMessage: React.Dispatch<React.SetStateAction<string>>;
-}) => {
-  const { t } = useTranslation('model');
-
-  const handleInput = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    e.target.style.height = 'auto';
-    e.target.style.height = `${e.target.scrollHeight}px`;
-    e.target.style.maxHeight = `${e.target.scrollHeight}px`;
-  };
-
-  const handleOnFocus = (e: React.FocusEvent<HTMLTextAreaElement, Element>) => {
-    e.target.style.height = 'auto';
-    e.target.style.height = `${e.target.scrollHeight}px`;
-    e.target.style.maxHeight = `${e.target.scrollHeight}px`;
-  };
-
-  const handleOnBlur = (e: React.FocusEvent<HTMLTextAreaElement, Element>) => {
-    e.target.style.height = 'auto';
-    e.target.style.maxHeight = '2.5rem';
-  };
-
-  return (
-    <div>
-      <div className='block text-sm font-medium text-gray-900 dark:text-white'>
-        {t('defaultSystemMessage')}
-      </div>
-      <textarea
-        className='my-2 mx-0 px-2 resize-none rounded-lg bg-transparent overflow-y-hidden leading-7 p-1 border border-gray-400/50 focus:ring-1 focus:ring-blue w-full max-h-10 transition-all'
-        onFocus={handleOnFocus}
-        onBlur={handleOnBlur}
-        onChange={(e) => {
-          _setSystemMessage(e.target.value);
-        }}
-        onInput={handleInput}
-        value={_systemMessage}
-        rows={1}
-      ></textarea>
-    </div>
   );
 };
